+++ addons/br/nfe/invoices.go
@@ -4,10 +4,10 @@ import (
 	"fmt"
 	"regexp"
 
+	"github.com/invopop/gobl/addons/br/dfe"
 	"github.com/invopop/gobl/bill"
 	"github.com/invopop/gobl/num"
 	"github.com/invopop/gobl/org"
-	"github.com/invopop/gobl/regimes/br"
 	"github.com/invopop/gobl/tax"
 	"github.com/invopop/validation"
 )
@@ -92,7 +92,7 @@ func validateParty(value any) error {
 		),
 		validation.Field(&p.Ext,
 			validation.When(len(p.Addresses) > 0,
-				tax.ExtensionsRequire(br.ExtKeyMunicipality),
+				tax.ExtensionsRequire(dfe.ExtKeyMunicipality),
 			),
 			validation.Skip,
 		),
@@ -107,7 +107,7 @@ func validateSupplier(value any) error {
 
 	return validation.ValidateStruct(s,
 		validation.Field(&s.Identities,
-			org.RequireIdentityKey(br.IdentityKeyStateReg),
+			org.RequireIdentityKey(dfe.IdentityKeyStateReg),
 			validation.Skip,
 		),
 		validation.Field(&s.Name, validation.Required),