+++ regimes/br/party_test.go
@@ -175,28 +175,3 @@ func TestValidatePostCodes(t *testing.T) {
 		})
 	}
 }
-
-func TestNormalizeParty(t *testing.T) {
-	t.Run("nil party", func(t *testing.T) {
-		var party *org.Party
-		br.Normalize(party)
-		assert.Nil(t, party)
-	})
-
-	t.Run("migrates old addon extension keys", func(t *testing.T) {
-		party := &org.Party{
-			Ext: tax.Extensions{
-				"br-nfse-fiscal-incentive": "1",
-				"br-nfse-municipality":     "1234567890",
-				"br-nfse-simples":          "2",
-				"br-nfse-special-regime":   "3",
-			},
-		}
-		br.Normalize(party)
-		assert.Len(t, party.Ext, 4)
-		assert.Equal(t, cbc.Code("1"), party.Ext["br-fiscal-incentive"])
-		assert.Equal(t, cbc.Code("1234567890"), party.Ext["br-municipality"])
-		assert.Equal(t, cbc.Code("2"), party.Ext["br-simples"])
-		assert.Equal(t, cbc.Code("3"), party.Ext["br-special-regime"])
-	})
-}